backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow>=10.2.0
numpy==1.26.3
python-dotenv==1.0.0
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
scikit-learn==1.4.0
scikit-image==0.22.0
matplotlib==3.8.2
httpx==0.26.0
twilio==8.11.0
sqlalchemy==2.0.25
